+++ constance/admin.py
@@ -17,7 +17,9 @@ from django.template.context import RequestContext
 from django.utils import six
 from django.utils.encoding import smart_bytes
 from django.utils.formats import localize
+from django.utils.module_loading import import_string
 from django.utils.translation import ugettext_lazy as _
+import django
 
 try:
     from django.utils.encoding import smart_bytes
@@ -48,11 +50,35 @@ FIELDS = {
     int: INTEGER_LIKE,
     Decimal: (fields.DecimalField, {'widget': NUMERIC_WIDGET}),
     str: STRING_LIKE,
-    datetime: (fields.DateTimeField, {'widget': widgets.AdminSplitDateTime}),
+    datetime: (fields.SplitDateTimeField, {'widget': widgets.AdminSplitDateTime}),
     date: (fields.DateField, {'widget': widgets.AdminDateWidget}),
     time: (fields.TimeField, {'widget': widgets.AdminTimeWidget}),
     float: (fields.FloatField, {'widget': NUMERIC_WIDGET}),
 }
+
+
+def parse_additional_fields(fields):
+    for key in fields:
+        field = list(fields[key])
+
+        if len(field) == 1:
+            field.append({})
+
+        field[0] = import_string(field[0])
+
+        if 'widget' in field[1]:
+            klass = import_string(field[1]['widget'])
+            field[1]['widget'] = klass(**(field[1].get('widget_kwargs', {}) or {}))
+
+            if 'widget_kwargs' in field[1]:
+                del field[1]['widget_kwargs']
+
+        fields[key] = field
+
+    return fields
+
+
+FIELDS.update(parse_additional_fields(settings.ADDITIONAL_FIELDS))
 
 if not six.PY3:
     FIELDS.update({
@@ -68,8 +94,14 @@ class ConstanceForm(forms.Form):
         super(ConstanceForm, self).__init__(*args, initial=initial, **kwargs)
         version_hash = hashlib.md5()
 
-        for name, (default, help_text, group) in settings.CONFIG.items():
-            config_type = type(default)
+        for name, options in settings.CONFIG.items():
+	    default, help_text = options[0], options[1]
+	    if len(options) == 4:
+               config_type = options[2]
+               group = options[3]
+            else:
+               config_type = type(default)		
+               group = options[2]
             if config_type not in FIELDS:
                 raise ImproperlyConfigured(_("Constance doesn't support "
                                              "config values of the type "
@@ -89,6 +121,10 @@ class ConstanceForm(forms.Form):
 
     def clean_version(self):
         value = self.cleaned_data['version']
+
+        if settings.IGNORE_ADMIN_VERSION_CHECK:
+            return value
+
         if value != self.initial['version']:
             raise forms.ValidationError(_('The settings have been modified '
                                           'by someone else. Please reload the '
@@ -98,6 +134,7 @@ class ConstanceForm(forms.Form):
 
 class ConstanceAdmin(admin.ModelAdmin):
     change_list_template = 'admin/constance/change_list.html'
+    change_list_form = ConstanceForm
 
     def get_urls(self):
         info = self.model._meta.app_label, self.model._meta.module_name
@@ -115,14 +152,14 @@ class ConstanceAdmin(admin.ModelAdmin):
         # First load a mapping between config name and default value
         if not self.has_change_permission(request, None):
             raise PermissionDenied
-        default_initial = ((name, default)
-            for name, (default, help_text, group) in settings.CONFIG.items())
+        default_initial = ((name, options[0])
+            for name, options in settings.CONFIG.items())
         # Then update the mapping with actually values from the backend
         initial = dict(default_initial,
             **dict(config._backend.mget(settings.CONFIG.keys())))
-        form = ConstanceForm(initial=initial)
+        form = self.change_list_form(initial=initial)
         if request.method == 'POST':
-            form = ConstanceForm(data=request.POST, initial=initial)
+            form = self.change_list_form(data=request.POST, initial=initial)
             if form.is_valid():
                 form.save()
                 # In django 1.5 this can be replaced with self.message_user
@@ -139,8 +176,10 @@ class ConstanceAdmin(admin.ModelAdmin):
             'opts': Config._meta,
             'form': form,
             'media': self.media + form.media,
+            'icon_type': 'gif' if VERSION < (1, 9) else 'svg',
         }
-        for name, (default, help_text, group) in settings.CONFIG.items():
+        for name, options in settings.CONFIG.items():
+            default, help_text = options[0], options[1]
             # First try to load the value from the actual backend
             value = initial.get(name)
             # Then if the returned value is None, get the default
@@ -150,7 +189,6 @@ class ConstanceAdmin(admin.ModelAdmin):
                 'name': name,
                 'default': localize(default),
                 'help_text': _(help_text),
-                'group': _(group),
                 'value': localize(value),
                 'modified': value != default,
                 'form_field': form[name],
@@ -163,7 +201,7 @@ class ConstanceAdmin(admin.ModelAdmin):
                                           current_app=self.admin_site.name)
         return TemplateResponse(request, self.change_list_template, context,
                                 **extra)
-        
+
     def has_add_permission(self, *args, **kwargs):
         return False
 
@@ -182,12 +220,11 @@ class Config(object):
         object_name = 'Config'
         model_name = module_name = 'config'
         verbose_name_plural = _('config')
-        get_ordered_objects = lambda x: False
         abstract = False
         swapped = False
 
-        # def get_ordered_objects(self):
-        #     return False
+        def get_ordered_objects(self):
+            return False
 
         def get_change_permission(self):
             return 'change_%s' % self.model_name