+++ tests/settings.py
@@ -51,6 +51,16 @@ long_value = 123456
 if not six.PY3:
     long_value = long(long_value)
 
+CONSTANCE_ADDITIONAL_FIELDS = {
+     'yes_no_null_select': ['django.forms.fields.ChoiceField',
+         {
+         'widget': 'django.forms.Select',
+         'choices': (("-----", None), ("yes", "Yes"), ("no", "No"))
+         }],
+     # note this intentionally uses a tuple so that we can test immutable
+     'email': ('django.forms.fields.EmailField',),
+}
+
 CONSTANCE_CONFIG = {
     'INT_VALUE': (1, 'some int'),
     'LONG_VALUE': (long_value, 'some looong int'),
@@ -63,7 +73,9 @@ CONSTANCE_CONFIG = {
     'FLOAT_VALUE': (3.1415926536, 'PI'),
     'DATE_VALUE': (date(2010, 12, 24), 'Merry Chrismas'),
     'TIME_VALUE': (time(23, 59, 59), 'And happy New Year'),
+    'CHOICE_VALUE': ('yes', 'select yes or no', 'yes_no_null_select'),
     'LINEBREAK_VALUE': ('Spam spam', 'eggs\neggs'),
+    'EMAIL_VALUE': ('test@example.com', 'An email', 'email'),
 }
 
 DEBUG = True