+++ tests/storage.py
@@ -27,6 +27,9 @@ class StorageTestsMixin(object):
         self.assertEqual(self.config.FLOAT_VALUE, 3.1415926536)
         self.assertEqual(self.config.DATE_VALUE, date(2010, 12, 24))
         self.assertEqual(self.config.TIME_VALUE, time(23, 59, 59))
+        self.assertEqual(self.config.CHOICE_VALUE, 'yes')
+        self.assertEqual(self.config.EMAIL_VALUE, 'test@example.com')
+
 
         # set values
         self.config.INT_VALUE = 100
@@ -39,6 +42,8 @@ class StorageTestsMixin(object):
         self.config.FLOAT_VALUE = 2.718281845905
         self.config.DATE_VALUE = date(2001, 12, 20)
         self.config.TIME_VALUE = time(1, 59, 0)
+        self.config.CHOICE_VALUE = 'no'
+        self.config.EMAIL_VALUE = 'foo@bar.com'
 
         # read again
         self.assertEqual(self.config.INT_VALUE, 100)
@@ -51,6 +56,9 @@ class StorageTestsMixin(object):
         self.assertEqual(self.config.FLOAT_VALUE, 2.718281845905)
         self.assertEqual(self.config.DATE_VALUE, date(2001, 12, 20))
         self.assertEqual(self.config.TIME_VALUE, time(1, 59, 0))
+        self.assertEqual(self.config.CHOICE_VALUE, 'no')
+        self.assertEqual(self.config.EMAIL_VALUE, 'foo@bar.com')
+
 
     def test_nonexistent(self):
         try: